Create mappings for new entities

+++ Theseus.Infrastructure/Dtos/ExamStepDto.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; } = default!;
         public ExamStageDto Stage { get; set; } = default!;
         public int Index { get; set; } = default!;
-        public byte StepTaken { get; set; } = default!;
+        public int StepTaken { get; set; } = default!;
         public float TimeBeforeStep { get; set; } = default!;
 
         public ExamStepDto() { }
